Creating new folder ( Header ) inside the components folder & moving HeaderLogo & HeaderNavbar into it
